feat: 게시글 수정/삭제는 작성자만 가능하도록 검증 추가
- 작성자가 아닐 경우 CustomAuthorizationException 하위 예외 발생

+++ src/main/java/ajaajaja/debugging_rounge/feature/question/application/QuestionService.java
@@ -1,5 +1,8 @@
 package ajaajaja.debugging_rounge.feature.question.application;
 
+import ajaajaja.debugging_rounge.common.exception.auth.CustomAuthorizationException;
+import ajaajaja.debugging_rounge.common.exception.auth.QuestionDeleteForbiddenException;
+import ajaajaja.debugging_rounge.common.exception.auth.QuestionUpdateForbiddenExceptionCustom;
 import ajaajaja.debugging_rounge.feature.question.api.dto.QuestionCreateRequestDto;
 import ajaajaja.debugging_rounge.feature.question.api.dto.QuestionDetailResponseDto;
 import ajaajaja.debugging_rounge.feature.question.api.dto.QuestionListResponseDto;
@@ -13,6 +16,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Objects;
+import java.util.function.Supplier;
 
 @Service
 @RequiredArgsConstructor
@@ -46,22 +52,29 @@ public class QuestionService {
     }
 
     @Transactional
-    public void updateQuestion(Long questionId, QuestionUpdateRequestDto questionUpdateRequestDto) {
-        Question question = questionRepository.findById(questionId).orElseThrow(QuestionNotFoundException::new);
+    public void updateQuestion(Long questionId, QuestionUpdateRequestDto questionUpdateRequestDto, Long loginUserId) {
+        Question question = questionRepository.findById(questionId)
+                .orElseThrow(QuestionNotFoundException::new);
+
+        validateAuthor(question.getUserId(), loginUserId, QuestionUpdateForbiddenExceptionCustom::new);
 
         question.update(questionUpdateRequestDto.getTitle(), questionUpdateRequestDto.getContent());
     }
 
     @Transactional
-    public void deleteQuestion(Long questionId) {
-        deleteQuestionOrThrow(questionId);
-    }
-
-    private void deleteQuestionOrThrow(Long questionId) {
+    public void deleteQuestion(Long questionId, Long loginUserId) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(QuestionNotFoundForDeleteException::new);
 
+        validateAuthor(question.getUserId(), loginUserId, QuestionDeleteForbiddenException::new);
+
         questionRepository.delete(question);
+    }
+
+    private void validateAuthor(Long authorId, Long loginUserId, Supplier<? extends CustomAuthorizationException> ex) {
+        if (!Objects.equals(authorId, loginUserId)) {
+            throw ex.get();
+        }
     }
 
 }
